+++ common/src/main/java/com/wynntils/screens/guides/gear/WynntilsItemGuideScreen.java
@@ -93,7 +93,7 @@ public final class WynntilsItemGuideScreen extends WynntilsGuideScreen<GuideGear
 
         return new GuideGearItemStackButton(
                 (int) (xOffset + Texture.CONTENT_BOOK_BACKGROUND.width() / 2f + 13 + offsetX),
-                (int) (yOffset + 43 + offsetY),
+                yOffset + 43 + offsetY,
                 18,
                 18,
                 elements.get(i),
