+++ common/src/main/java/com/wynntils/screens/guides/ingredient/WynntilsIngredientGuideScreen.java
@@ -123,7 +123,7 @@ public final class WynntilsIngredientGuideScreen
 
         return new GuideIngredientItemStackButton(
                 (int) (xOffset + Texture.CONTENT_BOOK_BACKGROUND.width() / 2f + 13 + offsetX),
-                (int) (yOffset + 43 + offsetY),
+                yOffset + 43 + offsetY,
                 18,
                 18,
                 elements.get(i),
